backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["jzhang1043 <jzhang1043@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
gunicorn = "^21.2.0"
pydantic = {extras = ["email"], version = "^2.6.4"}
bcrypt = "^4.1.2"
motor = "^3.4.0"
pydantic-settings = "^2.2.1"
beanie = "^1.25.0"
pymongo = "^4.6.3"
python-dotenv = "^1.0.1"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
